requirements.txt
pip >= 24.0.0

pylint==3.0.0
pytest==7.4.2
pytest-cov==4.1.0
black==23.9.1
#ipython==8.17.2

numpy
sounddevice
pyyaml

PyAudio
#adafruit-circuitpython-ssd1306
luma.core
luma.oled
Pillow

# spidev  ## install on Raspberry Pi
# RPi.GPIO  ## install on Raspberry Pi